Updating build script behaviour
Update the existing build scripts to take a build or clean parameter.
The former will just run the required "docker build" commands (without
running "docker rmi" first as previously). The second will delete the
local images in case a forced rebuild is required.

+++ gcp_deployment/build_images_chatui.sh
@@ -9,10 +9,11 @@ set -o pipefail  # don't hide errors within pipes
 # Arguments expected:
 #   $1 = path to full config file
 #   $2 = remote docker repo to tag images for
+#   $3 = either "build" to build+push images, or "clean" to remove all local images'
 
-if [[ $# -ne 2 ]]
+if [[ $# -ne 3 ]]
 then
-    echo "build_images_chatui.sh <config file path> <remote-repo>"
+    echo "build_images_chatui.sh <config file path> <remote-repo> <build/clean>"
     exit 1
 fi
 
@@ -21,14 +22,21 @@ declare -r remote="${2}"
 
 script_path="$( dirname -- "$0"; )"
 
-docker rmi -f chat:latest 2> /dev/null
-docker rmi -f "${remote}"/chat:latest 2> /dev/null
+if [[ "${3}" == "clean" ]]
+then
+    docker rmi -f chat:latest 2> /dev/null
+    docker rmi -f "${remote}"/chat:latest 2> /dev/null
+elif [[ "${3}" == "build" ]]
+then
+    pushd "${script_path}/../agent-dialogue-ui"
 
-pushd "${script_path}/../agent-dialogue-ui"
+    docker build -f Dockerfile -t chat:latest .
+    docker tag chat:latest "${remote}"/chat:latest
 
-docker build -f Dockerfile -t chat:latest .
-docker tag chat:latest "${remote}"/chat:latest
+    popd
 
-popd
-
-docker push "${remote}"/chat:latest
+    docker push "${remote}"/chat:latest
+else
+    echo "Unknown argument: ${3} (expected either 'build' or 'clean')"
+    exit 1
+fi
